# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_dao_model.UserAttribute:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      expiredAt:
        description: 过期时间
        type: string
      gid:
        description: GID
        type: string
      group:
        description: 课题组
        type: string
      id:
        description: ID
        type: integer
      name:
        description: 账号
        type: string
      nickname:
        description: 昵称，如果没有指定，则与账号相同
        type: string
      phone:
        description: 电话
        type: string
      teacher:
        description: 老师
        type: string
      uid:
        description: UID and GID are used for Filesystem
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_service_ResourceLimitCheckResult:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_service.ResourceLimitCheckResult'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-internal_handler_JobResourceSummaryResp:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/internal_handler.JobResourceSummaryResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_service.ResourceLimitCheckResult:
    properties:
      details:
        items:
          $ref: '#/definitions/github_com_raids-lab_crater_internal_service.ResourceLimitDetail'
        type: array
      enabled:
        type: boolean
      exceeded:
        type: boolean
    type: object
  github_com_raids-lab_crater_internal_service.ResourceLimitDetail:
    properties:
      exceeded:
        type: boolean
      limit:
        type: string
      resource:
        type: string
      used:
        type: string
    type: object
...
  internal_handler.JobResourceSummaryAcceleratorResp:
    properties:
      limit:
        type: string
      pending:
        type: string
      resource:
        type: string
      running:
        type: string
      used:
        type: string
    type: object
  internal_handler.JobResourceSummaryResp:
    properties:
      accelerators:
        items:
          $ref: '#/definitions/internal_handler.JobResourceSummaryAcceleratorResp'
        type: array
      cpu:
        $ref: '#/definitions/internal_handler.JobResourceSummaryUsageResp'
      memory:
        $ref: '#/definitions/internal_handler.JobResourceSummaryUsageResp'
      pendingJobs:
        type: integer
      runningJobs:
        type: integer
    type: object
  internal_handler.JobResourceSummaryUsageResp:
    properties:
      limit:
        type: string
      pending:
        type: string
      running:
        type: string
      used:
        type: string
    type: object
...
  internal_handler.ResourceLimitCheckReq:
    properties:
      requestedResources:
        additionalProperties:
          type: string
        type: object
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/context/attributes:
    put:
      consumes:
      - application/json
      description: Update the attributes of the current user
      parameters:
      - description: User attributes
        in: body
        name: attributes
        required: true
        schema:
          $ref: '#/definitions/github_com_raids-lab_crater_dao_model.UserAttribute'
      produces:
      - application/json
      responses:
        "200":
          description: User attributes updated
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Update user attributes
      tags:
      - Context
  /v1/context/email/code:
    post:
      consumes:
      - application/json
      description: generate random code and save, send it to the user's email
      produces:
      - application/json
      responses:
        "200":
          description: Successfully send email verification code to user
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Send User Verification Code for email
      tags:
      - Context
  /v1/context/email/update:
    post:
      consumes:
      - application/json
      description: judge code and update email for user
      produces:
      - application/json
      responses:
        "200":
          description: User email updated successfully
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Update after judging Verification Code for email
      tags:
      - Context
  /v1/context/job-resource-summary:
    get:
      consumes:
      - application/json
      description: 按个人视角汇总运行中与排队中的作业资源占用，并返回队列内资源限制
      produces:
      - application/json
      responses:
        "200":
          description: 当前作业资源占用汇总
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-internal_handler_JobResourceSummaryResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取当前资源占用汇总
      tags:
      - Context
  /v1/context/queue:
    get:
      consumes:
      - application/json
      description: query the queue information by client-go
      produces:
      - application/json
      responses:
        "200":
          description: Volcano Queue Quota
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Get the queue information
      tags:
      - Context
  /v1/context/resource-limit-check:
    post:
      consumes:
      - application/json
      description: 根据队列内资源限制配置，检查当前用户运行中作业资源加上本次请求资源是否超限
      parameters:
      - description: 本次请求的资源
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.ResourceLimitCheckReq'
      produces:
      - application/json
      responses:
        "200":
          description: 检查结果
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_service_ResourceLimitCheckResult'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 检查用户资源使用是否超限
      tags:
      - Context
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"